Short is too short, not long enough.
git-svn-id: https://everest.teltel.com:8443/repos/misc/9x9dev/trunk/nncloudtv@4609 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/main/java/com/nncloudtv/model/NnChannel.java
@@ -505,9 +505,9 @@ public class NnChannel implements Serializable {
             return 0;
         String[] splits = userId.split("-");
         if (splits.length > 1)
-            return Short.parseShort(splits[1]);
+            return Long.parseLong(splits[1]);
         else
-            return Short.parseShort(userId);        
+            return Long.parseLong(userId);        
     }
     
     public String getUserIdStr() {
